Full tests +small cleanup.
