+++ .mocharc.json
@@ -1,0 +1,5 @@
+{
+    "require": "hardhat/register",
+    "timeout": 40000,
+    "_": ["test/**/*.ts"]
+  }
